CDAP-14673 Refactor storage classes into co.cask.cdap.spi.data package
