+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/runtime/schedule/queue/SqlJobQueueTableTest.java
@@ -20,10 +20,10 @@ import co.cask.cdap.api.dataset.lib.KeyValueTable;
 import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.data2.sql.PostgresSqlStructuredTableAdmin;
-import co.cask.cdap.data2.sql.RetryingSqlTransactionRunner;
-import co.cask.cdap.data2.sql.SqlStructuredTableRegistry;
 import co.cask.cdap.spi.data.StructuredTableAdmin;
+import co.cask.cdap.spi.data.sql.PostgresSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.sql.RetryingSqlTransactionRunner;
+import co.cask.cdap.spi.data.sql.SqlStructuredTableRegistry;
 import co.cask.cdap.spi.data.transaction.TransactionRunner;
 import co.cask.cdap.store.StoreDefinition;
 import com.google.common.base.Joiner;
