+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/runtime/schedule/store/SqlProgramScheduleStoreDatasetTest.java
@@ -20,11 +20,11 @@ import co.cask.cdap.api.dataset.lib.KeyValueTable;
 import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.data2.sql.PostgresSqlStructuredTableAdmin;
-import co.cask.cdap.data2.sql.SqlStructuredTableRegistry;
-import co.cask.cdap.data2.sql.SqlTransactionRunner;
 import co.cask.cdap.spi.data.StructuredTableAdmin;
 import co.cask.cdap.spi.data.TableAlreadyExistsException;
+import co.cask.cdap.spi.data.sql.PostgresSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.sql.SqlStructuredTableRegistry;
+import co.cask.cdap.spi.data.sql.SqlTransactionRunner;
 import co.cask.cdap.spi.data.transaction.TransactionRunner;
 import co.cask.cdap.store.StoreDefinition;
 import com.google.common.base.Joiner;
