+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/profile/SqlProfileServiceTest.java
@@ -21,13 +21,13 @@ import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.api.metrics.MetricsSystemClient;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.data2.sql.PostgresSqlStructuredTableAdmin;
-import co.cask.cdap.data2.sql.SqlStructuredTableRegistry;
-import co.cask.cdap.data2.sql.SqlTransactionRunner;
 import co.cask.cdap.internal.AppFabricTestHelper;
 import co.cask.cdap.internal.app.store.DefaultStore;
 import co.cask.cdap.spi.data.StructuredTableAdmin;
 import co.cask.cdap.spi.data.TableAlreadyExistsException;
+import co.cask.cdap.spi.data.sql.PostgresSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.sql.SqlStructuredTableRegistry;
+import co.cask.cdap.spi.data.sql.SqlTransactionRunner;
 import co.cask.cdap.spi.data.transaction.TransactionRunner;
 import co.cask.cdap.store.StoreDefinition;
 import com.google.common.base.Joiner;
