+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/runtime/StorageModule.java
@@ -20,13 +20,13 @@ package co.cask.cdap.data.runtime;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.conf.SConfiguration;
-import co.cask.cdap.data2.nosql.NoSqlStructuredTableAdmin;
-import co.cask.cdap.data2.nosql.NoSqlStructuredTableRegistry;
-import co.cask.cdap.data2.nosql.NoSqlTransactionRunner;
-import co.cask.cdap.data2.sql.PostgresSqlStructuredTableAdmin;
-import co.cask.cdap.data2.sql.SqlStructuredTableRegistry;
-import co.cask.cdap.data2.sql.SqlTransactionRunner;
 import co.cask.cdap.spi.data.StructuredTableAdmin;
+import co.cask.cdap.spi.data.nosql.NoSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.nosql.NoSqlStructuredTableRegistry;
+import co.cask.cdap.spi.data.nosql.NoSqlTransactionRunner;
+import co.cask.cdap.spi.data.sql.PostgresSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.sql.SqlStructuredTableRegistry;
+import co.cask.cdap.spi.data.sql.SqlTransactionRunner;
 import co.cask.cdap.spi.data.table.StructuredTableRegistry;
 import co.cask.cdap.spi.data.transaction.TransactionRunner;
 import com.google.inject.Inject;
