+++ cdap-data-fabric/src/main/java/co/cask/cdap/spi/data/nosql/NoSqlStructuredRow.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.data2.nosql;
+package co.cask.cdap.spi.data.nosql;
 
 import co.cask.cdap.api.dataset.table.Row;
 import co.cask.cdap.data2.dataset2.lib.table.MDSKey;
