+++ cdap-data-fabric/src/main/java/co/cask/cdap/spi/data/nosql/NoSqlStructuredTableAdmin.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.data2.nosql;
+package co.cask.cdap.spi.data.nosql;
 
 import co.cask.cdap.api.dataset.DatasetAdmin;
 import co.cask.cdap.api.dataset.DatasetContext;
