+++ cdap-data-fabric/src/main/java/co/cask/cdap/spi/data/sql/PostgresSqlStructuredTableAdmin.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.data2.sql;
+package co.cask.cdap.spi.data.sql;
 
 import co.cask.cdap.spi.data.StructuredTableAdmin;
 import co.cask.cdap.spi.data.TableAlreadyExistsException;
