+++ cdap-master/src/test/java/co/cask/cdap/master/environment/k8s/MasterServiceMainTestBase.java
@@ -22,9 +22,9 @@ import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.guice.ConfigModule;
 import co.cask.cdap.data.runtime.SystemDatasetRuntimeModule;
 import co.cask.cdap.data2.dataset2.lib.table.leveldb.LevelDBTableService;
-import co.cask.cdap.data2.nosql.NoSqlStructuredTableAdmin;
-import co.cask.cdap.data2.nosql.NoSqlStructuredTableRegistry;
 import co.cask.cdap.spi.data.StructuredTableAdmin;
+import co.cask.cdap.spi.data.nosql.NoSqlStructuredTableAdmin;
+import co.cask.cdap.spi.data.nosql.NoSqlStructuredTableRegistry;
 import co.cask.cdap.spi.data.table.StructuredTableRegistry;
 import co.cask.cdap.store.StoreDefinition;
 import com.google.common.collect.Lists;
